+++ packages/rocketh-node/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rocketh/node
 
+## 0.18.8
+
+### Patch Changes
+
+- environment refactor for simpler extensions
+- Updated dependencies
+  - @rocketh/core@0.18.4
+  - rocketh@0.18.7
+
 ## 0.18.7
 
 ### Patch Changes