+++ packages/rocketh-proxy/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @rocketh/proxy
 
+## 0.18.7
+
+### Patch Changes
+
+- environment refactor for simpler extensions
+- Updated dependencies
+  - @rocketh/read-execute@0.18.5
+  - @rocketh/deploy@0.18.4
+  - @rocketh/core@0.18.4
+
 ## 0.18.6
 
 ### Patch Changes