+++ packages/rocketh-web/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rocketh/web
 
+## 0.18.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @rocketh/core@0.18.4
+  - rocketh@0.18.7
+
 ## 0.18.7
 
 ### Patch Changes